requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.17
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.2
django-jazzmin==3.0.1
django-js-asset==2.2.0
pillow==11.0.0
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
